+++ packages/kit/src/runtime/client/client.js
@@ -218,7 +218,16 @@ export function create_client({ target, base, trailing_slash }) {
 			throw new Error(`Attempted to prefetch a URL that does not belong to this app: ${url}`);
 		}
 
-		load_cache = { id: intent.id, promise: load_route(intent) };
+		load_cache = {
+			id: intent.id,
+			promise: load_route(intent).then((result) => {
+				if (result.type === 'loaded' && result.state.error) {
+					// Don't cache errors, because they might be transient
+					load_cache = null;
+				}
+				return result;
+			})
+		};
 
 		return load_cache.promise;
 	}
@@ -315,29 +324,8 @@ export function create_client({ target, base, trailing_slash }) {
 		if (opts) {
 			const { scroll, keepfocus } = opts;
 
-			if (!keepfocus) {
-				// Reset page selection and focus
-				// We try to mimic browsers' behaviour as closely as possible by targeting the
-				// first scrollable region, but unfortunately it's not a perfect match — e.g.
-				// shift-tabbing won't immediately cycle up from the end of the page on Chromium
-				// See https://html.spec.whatwg.org/multipage/interaction.html#get-the-focusable-area
-				const root = document.body;
-				const tabindex = root.getAttribute('tabindex');
-
-				root.tabIndex = -1;
-				root.focus({ preventScroll: true });
-
-				setTimeout(() => {
-					getSelection()?.removeAllRanges();
-				});
-
-				// restore `tabindex` as to prevent `root` from stealing input from elements
-				if (tabindex !== null) {
-					root.setAttribute('tabindex', tabindex);
-				} else {
-					root.removeAttribute('tabindex');
-				}
-			}
+			// reset focus first, so that manual focus management can override it
+			if (!keepfocus) reset_focus();
 
 			// need to render the DOM before we can scroll to the rendered elements
 			await tick();
@@ -356,7 +344,6 @@ export function create_client({ target, base, trailing_slash }) {
 				}
 			}
 		} else {
-			// in this case we're simply invalidating
 			await tick();
 		}
 
@@ -373,6 +360,8 @@ export function create_client({ target, base, trailing_slash }) {
 
 	/** @param {import('./types').NavigationFinished} result */
 	function initialize(result) {
+		if (__SVELTEKIT_DEV__ && document.querySelector('vite-error-overlay')) return;
+
 		current = result.state;
 
 		const style = document.querySelector('style[data-sveltekit]');
@@ -485,6 +474,12 @@ export function create_client({ target, base, trailing_slash }) {
 			};
 
 			// TODO remove this for 1.0
+			Object.defineProperty(result.props.page, 'routeId', {
+				get() {
+					throw new Error('$page.routeId has been replaced by $page.route.id');
+				},
+				enumerable: false
+			});
 			/**
 			 * @param {string} property
 			 * @param {string} replacement
@@ -599,8 +594,8 @@ export function create_client({ target, base, trailing_slash }) {
 
 					// prerendered pages may be served from any origin, so `initial_fetch` urls shouldn't be resolved
 					return started
-						? subsequent_fetch(resolved, init)
-						: initial_fetch(requested, resolved, init);
+						? subsequent_fetch(requested, resolved, init)
+						: initial_fetch(requested, init);
 				},
 				setHeaders: () => {}, // noop
 				depends,
@@ -1058,7 +1053,10 @@ export function create_client({ target, base, trailing_slash }) {
 			}
 		};
 
-		callbacks.before_navigate.forEach((fn) => fn(cancellable));
+		if (!navigating) {
+			// Don't run the event during redirects
+			callbacks.before_navigate.forEach((fn) => fn(cancellable));
+		}
 
 		return should_block ? null : navigation;
 	}
@@ -1287,6 +1285,8 @@ export function create_client({ target, base, trailing_slash }) {
 					const post_update = pre_update();
 					root.$set(navigation_result.props);
 					post_update();
+
+					tick().then(reset_focus);
 				}
 			} else if (result.type === 'redirect') {
 				goto(result.location, { invalidateAll: true }, []);
@@ -1299,6 +1299,10 @@ export function create_client({ target, base, trailing_slash }) {
 				const post_update = pre_update();
 				root.$set(props);
 				post_update();
+
+				if (result.type === 'success') {
+					tick().then(reset_focus);
+				}
 			}
 		},
 
@@ -1353,9 +1357,9 @@ export function create_client({ target, base, trailing_slash }) {
 
 			/** @param {Event} event */
 			const trigger_prefetch = (event) => {
-				const { url, options } = find_anchor(event);
-				if (url && options.prefetch) {
-					if (is_external_url(url)) return;
+				const { url, options, has } = find_anchor(event);
+				if (url && options.prefetch && !is_external_url(url)) {
+					if (options.reload || has.rel_external || has.target || has.download) return;
 					prefetch(url);
 				}
 			};
@@ -1387,7 +1391,7 @@ export function create_client({ target, base, trailing_slash }) {
 				if (event.metaKey || event.ctrlKey || event.shiftKey || event.altKey) return;
 				if (event.defaultPrevented) return;
 
-				const { a, url, options } = find_anchor(event);
+				const { a, url, options, has } = find_anchor(event);
 				if (!a || !url) return;
 
 				const is_svg_a_element = a instanceof SVGAElement;
@@ -1407,15 +1411,10 @@ export function create_client({ target, base, trailing_slash }) {
 				)
 					return;
 
-				if (a.hasAttribute('download')) return;
+				if (has.download) return;
 
 				// Ignore the following but fire beforeNavigate
-				const rel = (a.getAttribute('rel') || '').split(/\s+/);
-				if (
-					rel.includes('external') ||
-					options.reload ||
-					(is_svg_a_element ? a.target.baseVal : a.target)
-				) {
+				if (options.reload || has.rel_external || has.target) {
 					const navigation = before_navigate({ url, type: 'link' });
 					if (!navigation) {
 						event.preventDefault();
@@ -1678,6 +1677,36 @@ function pre_update() {
 	}
 
 	return () => {};
+}
+
+function reset_focus() {
+	const autofocus = document.querySelector('[autofocus]');
+	if (autofocus) {
+		// @ts-ignore
+		autofocus.focus();
+	} else {
+		// Reset page selection and focus
+		// We try to mimic browsers' behaviour as closely as possible by targeting the
+		// first scrollable region, but unfortunately it's not a perfect match — e.g.
+		// shift-tabbing won't immediately cycle up from the end of the page on Chromium
+		// See https://html.spec.whatwg.org/multipage/interaction.html#get-the-focusable-area
+		const root = document.body;
+		const tabindex = root.getAttribute('tabindex');
+
+		root.tabIndex = -1;
+		root.focus({ preventScroll: true });
+
+		setTimeout(() => {
+			getSelection()?.removeAllRanges();
+		});
+
+		// restore `tabindex` as to prevent `root` from stealing input from elements
+		if (tabindex !== null) {
+			root.setAttribute('tabindex', tabindex);
+		} else {
+			root.removeAttribute('tabindex');
+		}
+	}
 }
 
 if (__SVELTEKIT_DEV__) {