+++ packages/kit/test/apps/basics/src/routes/+layout.server.js
@@ -1,24 +1,15 @@
 import { error, redirect } from '@sveltejs/kit';
 
-let should_fail = '';
-/**
- * @param {string} value
- */
-export function set_should_fail(value) {
-	should_fail = value;
-}
-
 /** @type {import('./$types').LayoutServerLoad} */
-export async function load() {
+export async function load({ cookies }) {
+	const should_fail = cookies.get('fail-type');
 	if (should_fail) {
+		cookies.delete('fail-type', { path: '/' });
 		if (should_fail === 'expected') {
-			set_should_fail('');
 			throw error(401, 'Not allowed');
 		} else if (should_fail === 'unexpected') {
-			set_should_fail('');
 			throw new Error('Failed to load');
 		} else {
-			set_should_fail('');
 			throw redirect(307, '/load');
 		}
 	}