+++ packages/kit/test/apps/basics/test/test.js
@@ -5,79 +5,6 @@ import { start_server, test } from '../../../utils.js';
 
 test.describe.configure({ mode: 'parallel' });
 
-test.describe('a11y', () => {
-	test('resets focus', async ({ page, clicknav, browserName }) => {
-		const tab = browserName === 'webkit' ? 'Alt+Tab' : 'Tab';
-
-		await page.goto('/accessibility/a');
-
-		await clicknav('[href="/accessibility/b"]');
-		expect(await page.innerHTML('h1')).toBe('b');
-		expect(await page.evaluate(() => (document.activeElement || {}).nodeName)).toBe('BODY');
-		await page.keyboard.press(tab);
-
-		expect(await page.evaluate(() => (document.activeElement || {}).nodeName)).toBe('BUTTON');
-		expect(await page.evaluate(() => (document.activeElement || {}).textContent)).toBe('focus me');
-
-		await clicknav('[href="/accessibility/a"]');
-		expect(await page.innerHTML('h1')).toBe('a');
-		expect(await page.evaluate(() => (document.activeElement || {}).nodeName)).toBe('BODY');
-
-		await page.keyboard.press(tab);
-		expect(await page.evaluate(() => (document.activeElement || {}).nodeName)).toBe('BUTTON');
-		expect(await page.evaluate(() => (document.activeElement || {}).textContent)).toBe('focus me');
-
-		expect(await page.evaluate(() => document.documentElement.getAttribute('tabindex'))).toBe(null);
-	});
-
-	test('announces client-side navigation', async ({ page, clicknav, javaScriptEnabled }) => {
-		await page.goto('/accessibility/a');
-
-		const has_live_region = (await page.innerHTML('body')).includes('aria-live');
-
-		if (javaScriptEnabled) {
-			expect(has_live_region).toBeTruthy();
-
-			// live region should exist, but be empty
-			expect(await page.innerHTML('[aria-live]')).toBe('');
-
-			await clicknav('[href="/accessibility/b"]');
-			expect(await page.innerHTML('[aria-live]')).toBe('b'); // TODO i18n
-		} else {
-			expect(has_live_region).toBeFalsy();
-		}
-	});
-
-	test('reset selection', async ({ page, clicknav }) => {
-		await page.goto('/selection/a');
-
-		expect(
-			await page.evaluate(() => {
-				const range = document.createRange();
-				range.selectNodeContents(document.body);
-				const selection = getSelection();
-				if (selection) {
-					selection.removeAllRanges();
-					selection.addRange(range);
-					return selection.rangeCount;
-				}
-				return -1;
-			})
-		).toBe(1);
-
-		await clicknav('[href="/selection/b"]');
-		expect(
-			await page.evaluate(() => {
-				const selection = getSelection();
-				if (selection) {
-					return selection.rangeCount;
-				}
-				return -1;
-			})
-		).toBe(0);
-	});
-});
-
 test.describe('Imports', () => {
 	test('imports from node_modules', async ({ page, clicknav }) => {
 		await page.goto('/imports');
@@ -396,16 +323,41 @@ test.describe('Encoded paths', () => {
 		});
 	});
 
-	test('allows %-encoded characters in directory names', async ({ page, clicknav }) => {
-		await page.goto('/encoded');
-		await clicknav('[href="/encoded/$SVLT"]');
-		expect(await page.textContent('h1')).toBe('$SVLT');
-	});
-
-	test('allows %-encoded characters in filenames', async ({ page, clicknav }) => {
+	test('allows non-ASCII character in parameterized route segment', async ({ page, clicknav }) => {
 		await page.goto('/encoded');
 		await clicknav('[href="/encoded/@svelte"]');
 		expect(await page.textContent('h1')).toBe('@svelte');
+	});
+
+	test('allows characters to be represented as escape sequences', async ({ page, clicknav }) => {
+		await page.goto('/encoded/escape-sequences');
+
+		await clicknav('[href="/encoded/escape-sequences/:-)"]');
+		expect(await page.textContent('h1')).toBe(':-)');
+
+		await clicknav('[href="/encoded/escape-sequences/%23"]');
+		expect(await page.textContent('h1')).toBe('#');
+
+		await clicknav('[href="/encoded/escape-sequences/%2F"]');
+		expect(await page.textContent('h1')).toBe('/');
+
+		await clicknav('[href="/encoded/escape-sequences/%3f"]');
+		expect(await page.textContent('h1')).toBe('?');
+
+		await clicknav('[href="/encoded/escape-sequences/%25"]');
+		expect(await page.textContent('h1')).toBe('%');
+
+		await clicknav('[href="/encoded/escape-sequences/<"]');
+		expect(await page.textContent('h1')).toBe('<');
+
+		await clicknav('[href="/encoded/escape-sequences/1<2"]');
+		expect(await page.textContent('h1')).toBe('1<2');
+
+		await clicknav('[href="/encoded/escape-sequences/苗"]');
+		expect(await page.textContent('h1')).toBe('苗');
+
+		await clicknav('[href="/encoded/escape-sequences/🤪"]');
+		expect(await page.textContent('h1')).toBe('🤪');
 	});
 });