+++ packages/kit/test/apps/dev-only/src/routes/server-only-modules/dynamic-import/+page.svelte
@@ -1,7 +1,0 @@
-<script>
-	const mod = import('$lib/test.server.js');
-</script>
-
-{#await mod then resolved}
-	<p>{resolved.should_explode}</p>
-{/await}